Fix: merge vanilla compat in equipment
